# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express Node.js API with MongoDB",
    "description": "This is a sample API for an Express Node.js application",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3002/api/v1",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/products": {
      "get": {
        "summary": "Get all products",
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Create a new product",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewProduct"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid category or product data"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Get a product by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the product",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "summary": "Update a product by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the product",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProduct"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid category or product data"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "Delete a product by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the product",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "The product is completely deleted."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/users/get/count": {
      "get": {
        "summary": "Get total user count",
        "responses": {
          "200": {
            "description": "Total user count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "number",
                      "example": 10
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user data"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Invalid user data"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "User login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "user is authenticated"
                    },
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MDdlMGIyMTg5NDJhYzcwNDk5NjBhZmIiLCJpc0FkbWluIjpmYWxzZSwiaWF0IjoxNjYxOTY4NjY5LCJleHAiOjE2NjE5Njk4Njl9.k23RPfROa13ogFsyOv-Esx-8Oxfg04mFJZvfdiKbDxU"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "607ebcb31703622ef09e6097"
          },
          "name": {
            "type": "string",
            "example": "Product 1"
          },
          "description": {
            "type": "string",
            "example": "Description of Product 1"
          },
          "richDescription": {
            "type": "string",
            "example": "Rich description of Product 1"
          },
          "image": {
            "type": "string",
            "example": "product1.jpg"
          },
          "brand": {
            "type": "string",
            "example": "Brand 1"
          },
          "price": {
            "type": "number",
            "example": 99.99
          },
          "category": {
            "type": "string",
            "example": "607ebcb31703622ef09e6097"
          },
          "countInStock": {
            "type": "number",
            "example": 10
          },
          "rating": {
            "type": "number",
            "example": 4.5
          },
          "isFeatured": {
            "type": "boolean",
            "example": true
          },
          "dateCreated": {
            "type": "string",
            "format": "date-time",
            "example": "2022-04-21T11:23:10.760Z"
          }
        }
      },
      "NewProduct": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Product 1"
          },
          "description": {
            "type": "string",
            "example": "Description of Product 1"
          },
          "richDescription": {
            "type": "string",
            "example": "Rich description of Product 1"
          },
          "image": {
            "type": "string",
            "example": "product1.jpg"
          },
          "brand": {
            "type": "string",
            "example": "Brand 1"
          },
          "price": {
            "type": "number",
            "example": 99.99
          },
          "category": {
            "type": "string",
            "example": "607ebcb31703622ef09e6097"
          },
          "countInStock": {
            "type": "number",
            "example": 10
          },
          "rating": {
            "type": "number",
            "example": 4.5
          },
          "isFeatured": {
            "type": "boolean",
            "example": true
          },
          "dateCreated": {
            "type": "string",
            "format": "date-time",
            "example": "2022-04-21T11:23:10.760Z"
          }
        }
      },
      "UpdateProduct": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Product 1"
          },
          "description": {
            "type": "string",
            "example": "Description of Product 1"
          },
          "richDescription": {
            "type": "string",
            "example": "Rich description of Product 1"
          },
          "image": {
            "type": "string",
            "example": "product1.jpg"
          },
          "brand": {
            "type": "string",
            "example": "Brand 1"
          },
          "price": {
            "type": "number",
            "example": 99.99
          },
          "category": {
            "type": "string",
            "example": "607ebcb31703622ef09e6097"
          },
          "countInStock": {
            "type": "number",
            "example": 10
          },
          "rating": {
            "type": "number",
            "example": 4.5
          },
          "isFeatured": {
            "type": "boolean",
            "example": true
          },
          "dateCreated": {
            "type": "string",
            "format": "date-time",
            "example": "2022-04-21T11:23:10.760Z"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "607ebcb31703622ef09e6097"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "phone": {
            "type": "string",
            "example": "123-456-7890"
          },
          "country": {
            "type": "string",
            "example": "USA"
          }
        }
      },
      "NewUser": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "example": "password"
          },
          "phone": {
            "type": "string",
            "example": "123-456-7890"
          },
          "country": {
            "type": "string",
            "example": "USA"
          }
        }
      },
      "UserLogin": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "example": "password"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}